pyproject.toml
[project]
name = "codebase-embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "code-splitter>=0.1.5",
    "gitpython>=3.1.45",
    "openai>=2.6.1",
    "pinecone>=7.3.0",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "semantic-text-splitter>=0.28.0",
    "tree-sitter-hcl>=1.2.0",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
]
